requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.1.1
distro==1.9.0
Django==4.2.19
exceptiongroup==1.2.2
executing==2.2.0
frozenlist==1.5.0
genai==2.1.0
google-auth==2.38.0
google-genai==1.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipython==8.18.1
jedi==0.19.2
jiter==0.8.2
matplotlib-inline==0.1.7
multidict==6.1.0
openai==0.27.10
parso==0.8.4
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.2.1
psycopg2==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
tabulate==0.9.0
tiktoken==0.3.3
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
websockets==14.2
yarl==1.18.3
